Payload_Type/iris/requirements.txt
transformers = "4.37.0"
sentence-transformers = "2.2.2"
huggingface-hub = "0.20.3"
langchain = "0.1.2"
langchain-community = "0.0.14"
ctransformers >= "0.2.27",<"1.0.0"
mythic-container = "v0.4.14"
llama-index-embeddings-huggingface >= "0.1.4", <"1.0.0"
llama-index-llms-llama-cpp >= "0.1.3", <"1.0.0"
llama-index-readers-file >= "0.1.11", <"1.0.0"
llama-index-embeddings-openai >= "0.1.7", <"1.0.0"
llama-index-embeddings-langchain >= "0.1.2", <"1.0.0"
llama-index-llms-openai >= "0.1.12", <"1.0.0"
llama-index-readers-graphql >= "0.1.3", <"1.0.0"